+++ embedded-io-adapters/src/lib.rs
@@ -7,11 +7,25 @@
 // We don't immediately remove them to not immediately break older nightlies.
 // When all features are stable, we'll remove them.
 #![cfg_attr(
-    all(any(feature = "tokio-1", feature = "futures-03"), nightly),
+    all(
+        any(
+            feature = "tokio-1",
+            feature = "futures-03",
+            feature = "embedded-storage-async"
+        ),
+        nightly
+    ),
     allow(stable_features)
 )]
 #![cfg_attr(
-    all(any(feature = "tokio-1", feature = "futures-03"), nightly),
+    all(
+        any(
+            feature = "tokio-1",
+            feature = "futures-03",
+            feature = "embedded-storage-async"
+        ),
+        nightly
+    ),
     feature(async_fn_in_trait, impl_trait_projections)
 )]
 
@@ -26,3 +40,11 @@ pub mod futures_03;
 #[cfg(feature = "tokio-1")]
 #[cfg_attr(docsrs, doc(cfg(feature = "tokio-1")))]
 pub mod tokio_1;
+
+#[cfg(feature = "embedded-storage")]
+#[cfg_attr(docsrs, doc(cfg(feature = "embedded-storage")))]
+pub mod embedded_storage;
+
+#[cfg(feature = "embedded-storage-async")]
+#[cfg_attr(docsrs, doc(cfg(feature = "embedded-storage-async")))]
+pub mod embedded_storage_async;